feat:created labelers table for clients
